+++ src/main/java/net/yxiao233/ifeu/mixin/entity/MixinSimulatedHydroponicBedTile.java
@@ -1,65 +1,50 @@
 package net.yxiao233.ifeu.mixin.entity;
 
-import com.buuz135.industrial.api.plant.PlantRecollectable;
-import com.buuz135.industrial.block.agriculturehusbandry.tile.HydroponicBedTile;
+import com.buuz135.industrial.block.agriculturehusbandry.tile.SimulatedHydroponicBedTile;
 import com.buuz135.industrial.block.tile.IndustrialWorkingTile;
-import com.buuz135.industrial.item.HydroponicSimulationProcessorItem;
-import com.buuz135.industrial.utils.IFAttachments;
-import com.hrznstudio.titanium.component.progress.ProgressBarComponent;
+import com.hrznstudio.titanium.component.inventory.SidedInventoryComponent;
 import com.hrznstudio.titanium.module.BlockWithTile;
 import net.minecraft.core.BlockPos;
-import net.minecraft.core.Direction;
-import net.minecraft.nbt.CompoundTag;
-import net.minecraft.world.item.BlockItem;
-import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
-import net.minecraft.world.level.Level;
-import net.minecraft.world.level.block.Block;
-import net.minecraft.world.level.block.BushBlock;
 import net.minecraft.world.level.block.state.BlockState;
-import net.neoforged.neoforge.common.SpecialPlantable;
-import net.neoforged.neoforge.items.IItemHandler;
 import net.neoforged.neoforge.items.ItemHandlerHelper;
 import net.yxiao233.ifeu.api.item.IFEUAddonItem;
 import net.yxiao233.ifeu.api.item.IFEUAugmentTypes;
 import net.yxiao233.ifeu.common.utils.AugmentInventoryHelper;
+import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
 
 import java.util.ArrayList;
-import java.util.Iterator;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.function.Supplier;
 
-@Mixin(HydroponicBedTile.class)
-public abstract class MixinHydroponicBedTile extends IndustrialWorkingTile<HydroponicBedTile>{
+@Mixin(SimulatedHydroponicBedTile.class)
+public abstract class MixinSimulatedHydroponicBedTile extends IndustrialWorkingTile<SimulatedHydroponicBedTile> {
+    @Shadow @Final private SidedInventoryComponent<SimulatedHydroponicBedTile> output;
 
-    public MixinHydroponicBedTile(BlockWithTile basicTileBlock, int estimatedPower, BlockPos blockPos, BlockState blockState) {
+    public MixinSimulatedHydroponicBedTile(BlockWithTile basicTileBlock, int estimatedPower, BlockPos blockPos, BlockState blockState) {
         super(basicTileBlock, estimatedPower, blockPos, blockState);
     }
 
     @Inject(
-            method = "tryToHarvestAndReplant",
+            method = "work",
             at = @At(
                     value = "INVOKE",
                     target = "Ljava/util/List;forEach(Ljava/util/function/Consumer;)V"
             ),
             locals = LocalCapture.CAPTURE_FAILHARD
     )
-    private static void onForEachDrops(
-            Level level, BlockPos up, BlockState state, IItemHandler output, ProgressBarComponent<?> etherBuffer,
-            IndustrialWorkingTile<?> tile, Supplier<PlantRecollectable> plantSupplier, ItemStack simulationOutput,
-            CallbackInfoReturnable<Boolean> cir, PlantRecollectable cachedRecollectable, List<ItemStack> drops
-    ) {
+    public void onWork(CallbackInfoReturnable<IndustrialWorkingTile<SimulatedHydroponicBedTile>.WorkAction> cir, ItemStack simulationProcessor, ItemStack seed, ItemStack crop, long executions, double efficiency, List<ItemStack> generatedDrops){
         AtomicInteger apple = new AtomicInteger(0);
         List<ItemStack> list = new ArrayList<>();
-        int tier = AugmentInventoryHelper.getAugmentTier(tile, IFEUAugmentTypes.APPLE);
-        drops.forEach(stack -> {
+        int tier = AugmentInventoryHelper.getAugmentTier(this, IFEUAugmentTypes.APPLE);
+        generatedDrops.forEach(stack -> {
             if(!stack.is(Items.APPLE)){
                 list.add(stack);
             }else{
@@ -73,7 +58,7 @@ public abstract class MixinHydroponicBedTile extends IndustrialWorkingTile<Hydro
         }
 
         list.forEach(stack ->{
-            ItemHandlerHelper.insertItem(output, stack, false);
+            ItemHandlerHelper.insertItem(this.output, stack, false);
         });
     }